Summary
For the threeAndFive method, I used a for loop to cycle through all the numbers less than 1000. From there, I checked for what numbers were multiples of three and five. If they were a multiple, it was added to the sum. For the evenFibSum, I has a while loop for num being less than 4 million. The initial sum is 2 and num1 is 1 and num2 is two. Also, the entire num variable is equal to zero. With this, my num was equal to the last two numbers combined and I changed the num1 and num2 to the new numbers they become. I also had to check if the num%0==0 and if its true then the num gets added onto the sum. I return that sum. For the largestPrimeFactor() method, I create a for loop of the size of the number. In that, I check the num, divide the num by i, and decrement i. Then, I return i which is the largest prime factor of the number.